fix broken preview

+++ src/pages/Post.jsx
@@ -26,21 +26,27 @@ export default function Post() {
     const deletePost = () => {
         appwriteService.deletePost(post.$id).then((status) => {
             if (status) {
-                appwriteService.deleteFile(post.featuredImage);
+                if (post.featuredImage) {
+                    appwriteService.deleteFile(post.featuredImage);
+                }
                 navigate("/");
             }
         });
     };
 
+    const previewUrl = appwriteService.getFileView(post?.featuredImage);
+
     return post ? (
         <div className="py-8">
             <Container>
                 <div className="w-full flex justify-center mb-4 relative border rounded-xl p-2">
-                    <img
-                        src={appwriteService.getFilePreview(post.featuredImage)}
-                        alt={post.title}
-                        className="rounded-xl"
-                    />
+                    {previewUrl && (
+                        <img
+                            src={previewUrl}
+                            alt={post.title}
+                            className="rounded-xl"
+                        />
+                    )}
 
                     {isAuthor && (
                         <div className="absolute right-6 top-6">
